apps/services/nlp-parser/requirements.txt
fastapi
uvicorn
ingredient-parser-nlp
pydantic
typing